[FEAT] #25 프로필에서 반경설정 화면 삭제
