+++ app/src/main/res/navigation/nav_graph_register.xml
@@ -19,9 +19,6 @@
         android:name="com.starters.hsge.presentation.register.fragment.UserLocationFragment"
         android:label="UserLocationFragment"
         tools:layout="@layout/fragment_user_location">
-        <action
-            android:id="@+id/action_userLocationFragment_to_userDistanceFragment"
-            app:destination="@id/userDistanceFragment" />
     </fragment>
     <fragment
         android:id="@+id/userImageFragment"
@@ -103,9 +100,5 @@
             android:id="@+id/action_dogDislikeTagFragment_to_userLocationFragment"
             app:destination="@id/userLocationFragment" />
     </fragment>
-    <fragment
-        android:id="@+id/userDistanceFragment"
-        android:name="com.starters.hsge.presentation.register.fragment.UserDistanceFragment"
-        android:label="UserDistanceFragment"
-        tools:layout="@layout/fragment_user_distance"/>
+
 </navigation>
